chore(docs): add doc for the test server
